+++ mln_bot/lib/data.dart
@@ -1,2 +1,3 @@
-export "src/data/item_info.dart";
-export "src/data/utils.dart";
+export "src/data/item_utils.dart";
+export "src/data/message_utils.dart";
+export "src/data/user_utils.dart";
