+++ mln_bot/lib/src/commands/random_user.dart
@@ -1,5 +1,3 @@
-import "package:nyxx_commands/nyxx_commands.dart";
-
 import "utils.dart";
 
 final randomUserCommand = ChatCommand("random", "Find a random user", _randomUser);
