+++ mln_bot/lib/src/commands/utils.dart
@@ -1,14 +1,12 @@
-import "package:mln_bot/data.dart";
-import "package:mln_bot/secrets.dart";
-import "package:mln_bot/services.dart";
-import "package:mln_shared/data.dart" as mln;
-import "package:mln_shared/clients.dart";
-import "package:mln_shared/utils.dart";
+export "package:nyxx/nyxx.dart" hide Attachment, Cache, Message, User;
+export "package:nyxx_commands/nyxx_commands.dart";
 
-import "package:nyxx/nyxx.dart";
-import "package:nyxx_commands/nyxx_commands.dart";
-
+export "package:mln_bot/data.dart";
+export "package:mln_bot/secrets.dart";
+export "package:mln_bot/services.dart";
 export "package:mln_shared/mln_shared.dart";
+
+import "";
 
 Future<String?> checkUsername(String username) async {
   if (!username.startsWith("<@")) return username;
@@ -57,7 +55,7 @@ extension ChatUtils on ChatContext {
     ),
   );
 
-  Future<void> respondUser(mln.User user, String prefix) =>
+  Future<void> respondUser(User user, String prefix) =>
     respond(user.describe(prefix));
 
   Future<void> respondLogin({bool promptToRetry = false}) async {
