+++ mln_bot/lib/src/commands/whois.dart
@@ -1,5 +1,3 @@
-import "package:nyxx_commands/nyxx_commands.dart";
-
 import "utils.dart";
 
 final userQuery = ChatCommand(
