+++ mln_shared/lib/data.dart
@@ -1,5 +1,6 @@
 export "src/data/attachment.dart";
 export "src/data/badge.dart";
 export "src/data/friendship.dart";
+export "src/data/item_info.dart";
 export "src/data/message.dart";
 export "src/data/user.dart";
